# [6.5.0] Restart at most once when prepopulating repository rule environment